v2: change dot output
This is a first and partial try at improving the dot representation
of our pdag. It has a couple of subtle issues and needs more work.
The core idea is to represent parsers with additional boxes, what
permits us to give a better indication of the flow of control we
see. The current implementation is better than the "not at all"
we had so far, but it still has problems with equal-length
"alternative" sequences. Also, user-defined types are not yet
properly displayed.

see also https://github.com/rsyslog/liblognorm/issues/114